requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
requests
faster-whisper==1.2.0
whisperx==3.7.4
pyannote.audio==3.4.0
transformers[sentencepiece]==4.34.0
huggingface-hub==0.35.3
torch==2.8.0
